fix(all): processing algorithms for loading data and preparing the application for operation
base.py: the new YassCore base class automatically creates factories for actors and services and registers them; the ContextMixin and InitialContext classes are excluded due to the localization of the context in the EnvBuilder class
repo_utils_io.py: There are new handler classes for processing read and write streams, as well as functions for dispatching methods of these classes depending on the input and output format, respectively.
repo.py: the logic of selecting and creating memory buffers is hidden behind the RepoBuffer descriptor class, which will reduce the amount of code in Repo classes and unify the reading and writing procedure; a data deduplication procedure has been introduced, which is performed before uploading them to the backend.
schemas.py: the assignment of the circuit class and its design have been redesigned since the use of Polaris in the project. The json schema is now parsed based on jsonpath instead of jmespath, and the schema class is now associated with read and write controllers (see file repo_utils_io.py ).
scrapers.py: the logic of stopping the traversal of a part of the resource has become more explicit and is associated with read and write controllers (corresponding attributes have appeared). Also, now Scrapers themselves directly process the content of the resource before sending the data to the Repo.
yass.py: now the application class loads all the necessary modules by itself to register factory methods and classes. In addition, the application class has become a factory class and it is endowed with a method of creating contexts that it registers in the internal data structure.
other: processing with linters, removing unused lines of code, etc.