feat(resources): complete package refactoring
The interfaces of all classes have been significantly expanded.
The resource is now a global repository that defines the top-level
parameters of calls to the data source and stores instances of requests
to the resource.
In addition to the request fields, the resource request class now also
stores an activity trigger. The package introduces a new type of trigger
- resource end trigger. It allows data collectors to know whether the
data flow has ended during the next iteration of the data source.