feat(storages)!: full package modify
The interface of objects such as storages has been significantly changed
and expanded, and the basic components of storages have been expanded.
A new ContentQueue class has been introduced, which allows share storage
buffer memory between data collectors and write to it competitively and
independently. Now any write operation to the store is accompanied by
the creation of a ContentQueue instance. All storages now have a limit
that allows you to regulate the number of objects stored in memory
before they are loaded into the backend.
For each storage engine, its own instance of the storage class is now
created, thereby defining the role of this type of object as a proxy
over the real backend.