feat(app)!: app methods extended
The main class now also instantiates data stores. The main class also
tries to resolve the implementation of the event loop - first of all it
will try to use uvloop. The application's event loop runs in a separate
thread to avoid situations where the main thread is already blocked by
a third-party application at the start.