chore: changes in docs and project details
The previously used Yass project name turned out to be reserved, and
therefore changes were made to all parts of the project where it was
used, mainly in the documentation and import instructions in modules.
In addition, minor changes have been made, namely to the project build
parameters, git hook parameters, and the py.typed file has been added
for compatibility with PEP561.